+++ include/Math/CameraMath.h
@@ -84,7 +84,7 @@ public:
 	 * \returns
 	 * point coordinates
 	 */
-	static osg::Vec3d getPointOnVector(osg::Vec3d p1, osg::Vec3d p2, float distance);
+	static osg::Vec3d getPointOnVector(osg::Vec3d p1, osg::Vec3d p2, double distance);
 
 	/*!
 	 * \brief
