+++ src/Layout/Shape_CylinderSurface.cpp
@@ -38,7 +38,7 @@ osg::Vec3f Shape_CylinderSurface::secondBaseCenter() const
 
 float Shape_CylinderSurface::baseRadius() const
 {
-	return (mBaseRadiusSlider.value() / 100.0) * 250;
+	return ((float) mBaseRadiusSlider.value() / 100.0f) * 250.f;
 }
 
 }
