+++ src/OsgQtBrowser/QGraphicsViewAdapter.cpp
@@ -73,8 +73,8 @@ QGraphicsViewAdapter::QGraphicsViewAdapter(osg::Image* image, QWidget* widget):
 #endif
 	_graphicsView->viewport()->setParent(0);
 
-	int width = _graphicsScene->width();
-	int height = _graphicsScene->height();
+	int width = (int) _graphicsScene->width();
+	int height = (int) _graphicsScene->height();
 
 	_qimages[0] = QImage(QSize(width, height), QImage::Format_ARGB32);
 	_qimages[0].fill(_backgroundColor.rgba());
@@ -262,7 +262,7 @@ bool QGraphicsViewAdapter::handlePointerEvent(int x, int y, int buttonMask)
 {
 	osg::notify(osg::INFO)<<"dispatchPointerEvent("<<x<<", "<<y<<", "<<buttonMask<<")"<<std::endl;
 
-	y = _graphicsScene->height()-y;
+	y = (int) (_graphicsScene->height()-y);
 
 	bool leftButtonPressed = (buttonMask & osgGA::GUIEventAdapter::LEFT_MOUSE_BUTTON)!=0;
 	bool middleButtonPressed = (buttonMask & osgGA::GUIEventAdapter::MIDDLE_MOUSE_BUTTON)!=0;
